api_gateway/requirements.txt
bcrypt==3.2.2
cryptography==37.0.2
email-validator==1.2.1
fastapi==0.78.0
grpcio==1.46.3
grpcio-tools==1.46.3
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.17.6
